+++ collector/src/main/java/com/bitmovin/analytics/persistence/PersistingAuthenticatedDispatcher.kt
@@ -8,6 +8,7 @@ import com.bitmovin.analytics.data.BackendFactory
 import com.bitmovin.analytics.data.CacheConsumingBackend
 import com.bitmovin.analytics.data.EventData
 import com.bitmovin.analytics.data.IEventDataDispatcher
+import com.bitmovin.analytics.data.SEQUENCE_NUMBER_LIMIT
 import com.bitmovin.analytics.license.AuthenticationCallback
 import com.bitmovin.analytics.license.AuthenticationResponse
 import com.bitmovin.analytics.license.LicenseCall
@@ -90,6 +91,11 @@ internal class PersistingAuthenticatedDispatcher(
     }
 
     override fun add(data: EventData) {
+        // Do not send events with sequence number greater than the limit
+        if (sampleSequenceNumber > SEQUENCE_NUMBER_LIMIT) {
+            return
+        }
+
         data.sequenceNumber = sampleSequenceNumber++
 
         when (operationMode) {